Fix actual_width calculation
When calculating `padding_width`, we divide the
`remaining_padding_per_buffer` by the `SIDES_OF_BUFFER` (2) in order to
determine the _real_ length of the padding. Then, we `math.min()` the
user's specified `g:bufferline.maximum_padding` along with the
calculated `remaining_width_per_buffer`.

Then, to calculate `actual_width`, we _were_ performing this
calculation:

```lua
used_width + (padding_width * buffers_length)
```

However, this calculation forgets to take into account that, since we
have split the buffer padding in two, we now have to double it _again_ in
order to determine the true length of the buffers (since the padding is
duplicated; once on the left side, once on the right). So, to do this,
we instead perform this calculation:

```lua
used_width + (padding_width * buffers_length * SIDES_OF_BUFFER)
```

Now, padding on both sides of the buffer is accounted for. This
addresses issues that arise when the `g:bufferline.maximum_padding` is
set to a value that is always greater than `remaining_width_per_buffer`.
For example:

```vim
function! s:maxpadding()
  return float2nr(floor(&columns / 2))
endfunction

let g:bufferline = {
\  'animation': v:false,
\  'closable': v:true,
\  'clickable': v:false,
\  'icon_separator_active': '▊',
\  'icons': 'both',
\  'maximum_padding': s:maxpadding(),
\  'semantic_letters': v:true,
\}

" Keep tabs maxed
augroup barbar_padding
  autocmd!
  autocmd VimResized * let g:bufferline.maximum_padding = s:maxpadding()
augroup END
```