bug(render): no tabpages and BufferTabpageFill != TabLineFill
Whenever a user has both `BufferTabpageFill` _and_ `TabLineFill` set,
but they are both different, it will show both at the same time, _and_ a
separator using `BufferTabpageFill`.

The last change made it so that a separator would never be shown unless
there were tabpages— I don't think that's the right thing to do. I
changed it again so that separators could be shown, but the highlight
group is dependent on the layout having tabpages.