fix(bufferline_update): update_names on bufnew
Previously I had the buffer set to update on `BufNewFile`. However, the
issue still occurs if the `BufNew` is a preexisting entry in the
filesystem.

This commit is well out of the scope of this PR, but I didn't want to
give @romgrk a ton of work to review at once.

SEE #58