ref(render): consistently use BufferTabpageFill
Based on some discussion at this URL:

- romgrk/barbar.nvim/pull/64#discussion_r529081951

…it has been decided that we should use this group consistently.

That brings in to question the _name_ `BufferTabpageFill`, which
indicates that the 'Tabpage' is inherently part of the group. If I had
known that _this_ was going to become its legacy, I probably would have
chosen a better name for it during the implementation (like
`BufferLineFill`).

Howver, that bridge has been crossed.

SEE #58