ref(render): always show buffer index when available
Previously, the buffer index was hidden when entering jump mode (if the
buffer index was previously being shown on mode enter). Now, the buffer
index will stay in the foreground, and the buffer index will either take
the space of the first letter of the word, or the space of the file
icon.