fix(render): buffers with '%' in the name
This is a character that is integral to the syntax of the tabline
setting. To fix this, I had to escape all '%' in a buffer name with
'%%', which renders '%' literally.

Implementing this was a little challenging, as barbar previously
assumed that all characters in a given `group[2]` were visible
characters (i.e. characters that take up space on the screen). To get
around this, I added `group[3]`, which is the LITERAL string that is
shown on screen, while `group[2]` is now the string which is used in the
tabline in order to render `group[3]`.