fix(render): ensure buffer is clickable
This change makes it so that the following elements of the buffer are
always clickable if they are present:

1. The buffer separator,
2. The buffer icon,
3. The buffer name.

They were getting cut off by `slice_groups_left`, as the only place
where the main click handler was being placed was at the start of a
buffer item. So, when a buffer went off the side of the screen, it
became unclickable.