perf(jump_mode): split `letters` to char array
I believe this was originally intended, as `JumpMode.letters` was marked
as `-- array` in the code. However, it was actually a string and as such
we were doing `string.sub` and `strcharpart` to index it as if it were a
list.

This commit makes us pay the performance cost of splitting the string
all at once, and then the result is cached so that we can keep indexing
it to assign letters.