fix: lazy loading
This plugin has had some pretty persistent problems with lazy loading.

It stems from the fact that many users opt to overwriting
`vim.g.bufferline` themselves, and sometimes (due to variances in
configuration) this occurs _after_ our `plugin/bufferline.lua` file has
called the initial `bufferline.setup()` function. This means that we end
up with a `vim.g.bufferline` which has no default values, and thus when
we rightly assume that they will be present, errors start getting thrown
at the user.

To solve this, I created an `options` module which resolves the correct
value for each setting when it is accessed, rather than at `setup`. This
_does_ incur a bit of overhead, but I think it will be worth it in the
end if we can finally be rid of these issues!