fix(render): nvim-session-manager incompatability
Closes #312

Sidenote, but I am not convinced that `xolox/vim-session` is properly
integrated by the code in this `User` autocommand. I searched the repo,
and apparently the _only_ references to `session_save_commands` on GitHub
are from `barbar.nvim`. So where did it even come from?

We might want to look at removing it at some point.