fix(bufferline): ensure `:BufferClose`/`Delete` stay the same
When this feature (choosing the buffer to focus) was first implemented,
only the `nvim_create_command` for `:BufferClose` was updated. Thus,
even though `:BufferDelete` is supposed to be a synonym for
`:BufferClose`, the two commands did not do the same thing.

We use a `for` now to ensure that the two commands will always and
forever be equivalent.