fix(icons): being cleared by `:colorscheme`
The problem was that `nvim-web-devicons` resets its highlights using an
autocmd on `ColorScheme`, just to make sure that nothing weird happens
to the highlight groups it defines as a result of e.g. `:highlight
clear` when switching colorschemes. Our `ColorScheme` event was running
too soon (i.e. before the devicons event) and as such the icons were all
"cleared" upon calling `:colorscheme`.