perf(render): reuse `current_buffer_index` for scroll locking buffer
We already had a `current_buffer_index` variable in
`get_bufferline_containers`, so we can use that instead of storing all
the activities of each buffer in their containers and looping over the
buffers in each container looking for the current buffer,

Apologies for the style changes; I was debugging something else,
and had a really hard time reading this particular section. While
reading through to understand the changes I discovered this performance
improvement.