ref: granularize imports (#451)
* ref: granularize imports

* fix: sparse `icons.diagnostics` doesn't work

If you do:

```lua
require'barbar'.setup {
  icons = {
    diagnostics = {
      [vim.diagnostic.ERROR] = {enabled = true},
      [vim.diagnostic.HINT] = {enabled = true},
    },
  },
}
```

…it will not work. This is because a function in `Buffer` was using
`ipairs`, which stops upon reaching the first `nil` numerical index in a
`table`. However, upon fixing this, I discovered another bug: `icons`
was not being fully merged.

We had a snippet of code to "deep extend" `icons.diagnostics`, but since
we were only using `vim.tbl_deep_extend` on the other `icons` options,
they did not inherit the diagnostics. I extracted this snippet to a
function and called it on all `icons.modified`, `icons.current`, etc.

* ref: `'left'|'right'` → `side` alias

We reference distinct sides of the barbar often enough that it probably
deserves its own alias.

* perf(render): reuse `current_buffer_index` for scroll locking buffer

We already had a `current_buffer_index` variable in
`get_bufferline_containers`, so we can use that instead of storing all
the activities of each buffer in their containers and looping over the
buffers in each container looking for the current buffer,

Apologies for the style changes; I was debugging something else,
and had a really hard time reading this particular section. While
reading through to understand the changes I discovered this performance
improvement.

* perf(config): generate defaults inside `setup`

Removes unecessary `deepcopy`