feat: case-insentive sorting (#598)
* feat(config): `ignore_case` option

* feat(api): provide configured sort case converter in `with_pin_order`

This seems like the cleanest way to inject the conversion. We don't know
what is applicable to be converted from within the function, and so we
can't apply it automatically. On the other hand, if we make this a
separate helper function, it will be cumbersome to call it in each
sorting function that needs it.

I would be open to hearing other ideas, though.

* ref(api): use `to_sort_case` in sorting functions where applicable

* docs: `sort.ignore_case`

* perf(api): use function instead of closure

SEE: https://github.com/romgrk/barbar.nvim/pull/598#discussion_r1678639195