fix(events): respond to gitsigns update in remote buffers (#605)
It turns out that our gitsigns update autocmd was occasionally looking
in the wrong place for the number of the buffer that had its signs
updated.

It seems that it gitsigns may issue a secondary buffer number in the
user data event field when the buffer whose signs were updated is
different than the current buffer. I am not positive on the reason for
this, but I will assume that it was well-informed.