feat(barbar.Highlight): implement `resetup` to `setup` from scratch
It is common that we need to respond to an event which invalidates our
highlight cache and setup highlighting again without it. Previously, we
simply wrote these instructions out by hand:

```lua
barbar.utils.highlight.reset_cache()
barbar.highlight.setup()
```

However, as this operation is so common, it would be easier to
communicate that this happens in multiple places by extracting the
behavior into a function which can be searched for using LSP references.