fix(barbar.Events.enable): `barbar.Highlight.resetup` on `&bg` change
It seems that we had an oversight in determining when the highlight
cache needed to be invalidated. We listened for the colorscheme to be
swapped, but not for mutations to the current colorscheme.

One such mutation is `set bg=dark` vs `set bg=light`. This typically
does not emit a `ColorScheme` event, but still requires us to invalidate
the highlight cache and setup highlights.