Reset highlights when `&bg` changes (#626)
* feat(barbar.Highlight): implement `resetup` to `setup` from scratch

It is common that we need to respond to an event which invalidates our
highlight cache and setup highlighting again without it. Previously, we
simply wrote these instructions out by hand:

```lua
barbar.utils.highlight.reset_cache()
barbar.highlight.setup()
```

However, as this operation is so common, it would be easier to
communicate that this happens in multiple places by extracting the
behavior into a function which can be searched for using LSP references.

* ref(barbar.Events.enable): use `highlight.resetup`

This is one area where we had been manually wiping the highlight cache
and setting up highlights. Now that we have a function which does this,
we can do it inline & without defining a closure to do it.

* fix(barbar.Events.enable): `barbar.Highlight.resetup` on `&bg` change

It seems that we had an oversight in determining when the highlight
cache needed to be invalidated. We listened for the colorscheme to be
swapped, but not for mutations to the current colorscheme.

One such mutation is `set bg=dark` vs `set bg=light`. This typically
does not emit a `ColorScheme` event, but still requires us to invalidate
the highlight cache and setup highlights.