Normalize filepaths (#619)
* feat(fs): `absolute` path converter

This will be necessary when writing the ad-hoc `normalize`
implementation for Nvim 0.7

* feat(fs): `normalize` function

This is necessary because we support some versions of Neovim which do
not have `vim.fs` available. Essentially, we check if the version of
Neovim has `vim.fs`, and use it if possible. Otherwise, we use a limited
implementation.

* feat(fs): `split` path into parts

We perform this operation in several parts of the codebase. Here, we
define a single way to do this, and can update behavior from a single
location.

* feat(fs): filepath parts `join` operation

This is the inverse of `fs.split`. Since we offer the ability to create
parts of a filepath from the whole, we should also provide the ability
to recreate the whole from the parts.

It so happens that this functionality is necessary in several places
throughout the codebase.

* feat(fs): `slice_parts_from_end`

There is a common pattern in the code which does the following:

1. split some filepath
2. slice the end of a filepath to match a given depth
3. join the slice to form a new filepath

This function provides a utility to perform this function in one step.

* ref(api): use new `fs` filepath utilities

As we now have standard implementations of what these filepath
operations should be like in different versions of neovim, there is also
no need to branch inside of these functions to handle whether or not
`vim.fs` is available.

* ref(buffer): use new `fs` filepath utilities

We were able to reduce a bit of boilerplate here by relying on
`fs.slice_parts_from_end` rather than performing the split -> slice ->
join operations manually.